move columns and add table_name to column
