+++ seed/models/__init__.py
@@ -9,6 +9,7 @@
 # Import all the models in this folder
 
 # TODO: Should we move all the views to the, well, views folder?
+from .columns import *  # noqa
 from .cycles import *  # noqa
 from .models import *  # noqa
 from .properties import *  # noqa
